Loaded note data and displayed it in an editable text box, with temporary back button.

+++ app/src/main/java/com/notemd/storage/NoteDao.java
@@ -17,7 +17,7 @@ public interface NoteDao {
     Single<List<Note>> getAllNotes();
 
     @Query("SELECT * FROM note WHERE noteId = :noteId LIMIT 1")
-    Maybe<Note> getNote(int noteId);
+    Maybe<Note> getNote(long noteId);
 
     @Insert
     Single<Long> insertNote(Note note);
